+++ Main.java
@@ -10,16 +10,9 @@ public class Main {
         cart.addItem(new Item("jeans", 1, new BigDecimal("40.00"), "clothing"));
         cart.addItem(new Item("laptop", 1, new BigDecimal("999.99"), "electronics"));
 
-        // Construct receipt
-        System.out.println("--- Receipt ---");
-        cart.printReceiptLines();
-        System.out.println();
-        cart.printSubtotalLine();
-        cart.printDiscountsLine();
-        cart.printSubtotalAfterDiscountsLine();
-        cart.printTotalTaxLine();
-        cart.printGrandTotal();
-        System.out.println("-----------------");
-        System.out.println("Thank you for shopping!");
+        // Print receipt
+        ShoppingCartPrinter printCart = new ShoppingCartPrinter(cart);
+        printCart.printFullReceipt();
+
     }
 }